+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # create-llama
 
+## 0.1.39
+
+### Patch Changes
+
+- 3fb93c7: Use LlamaCloud pipeline for data ingestion in TS (private file uploads and generate script)
+
+## 0.1.38
+
+### Patch Changes
+
+- bd5e39a: Fix error that files in sub folders of 'data' are not displayed
+
 ## 0.1.37
 
 ### Patch Changes